chore: add scripts type file
